Merge
Additional adjustments and fixes necessary for the merge from workallocation-3.

+++ src/work-allocation/containers/task-home/task-home.component.ts
@@ -3,10 +3,9 @@ import { ActivatedRouteSnapshot, NavigationEnd, Router, RoutesRecognized } from 
 import { SubNavigation } from '@hmcts/rpx-xui-common-lib';
 
 import { AppUtils } from '../../../app/app-utils';
-import { InfoMessage, InfoMessageType } from '../../enums';
+import { InformationMessage } from '../../models/comms';
 import { TaskSortField } from '../../models/tasks';
 import { InfoMessageCommService } from '../../services';
-import { InformationMessage } from '../../models/comms';
 
 @Component({
   selector: 'exui-task-home',
@@ -34,6 +33,18 @@ export class TaskHomeComponent implements OnInit {
 
   public infoMessages: InformationMessage[];
 
+  /**
+   * Flag to indicate whether or not messages should be retained at
+   * this point, having been set on another route. The flag is passed
+   * through the router and so is held in window.history.state.
+   */
+  private get retainMessages(): boolean {
+    if (window && window.history && window.history.state) {
+      return !!window.history.state.retainMessages;
+    }
+    return false;
+  }
+
   constructor(
     private readonly router: Router,
     private readonly infoMessageCommService: InfoMessageCommService
@@ -44,6 +55,7 @@ export class TaskHomeComponent implements OnInit {
       if (event instanceof NavigationEnd) {
         // Set up the active navigation item.
         this.setupActiveSubNavigationItem(this.router.url);
+        this.resetMessages();
       }
       if (event instanceof RoutesRecognized) {
         // Set up the page data.
@@ -62,7 +74,6 @@ export class TaskHomeComponent implements OnInit {
   public subscribeToInfoMessageCommService(): void {
 
     this.infoMessageCommService.infoMessageChangeEmitted$.subscribe(messages => {
-
       this.infoMessages = messages;
       this.showInfoMessage = true;
     });
@@ -75,7 +86,7 @@ export class TaskHomeComponent implements OnInit {
   public setupActiveSubNavigationItem(activeUrl: string): void {
     if (this.subNavigationItems) {
       for (const item of this.subNavigationItems) {
-        item.active = item.href === activeUrl;
+        item.active = this.isActiveUrl(item.href, activeUrl);
       }
     }
   }
@@ -87,7 +98,20 @@ export class TaskHomeComponent implements OnInit {
   public setupPageData(activatedRoute: ActivatedRouteSnapshot): void {
     const data = AppUtils.getRouteData(activatedRoute);
     this.pageTitle = data ? data.subTitle : 'Task list';
+  }
 
-    this.infoMessageCommService.removeAllMessages();
+  public resetMessages(): void {
+    // Remove all the messages unless they've been specifically flagged
+    // to be retained this one time.
+    if (!this.retainMessages) {
+      this.infoMessageCommService.removeAllMessages();
+    }
+  }
+
+  public isActiveUrl(url: string, currentUrl: string): boolean {
+    if (url && currentUrl) {
+      return currentUrl.indexOf(url) === 0;
+    }
+    return false;
   }
 }
